% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchments.R
\name{catchments}
\alias{catchments}
\title{compute catchments}
\usage{
catchments(locs, pc)
}
\arguments{
\item{locs}{spatial coordinates}

\item{pc}{return catchment area at this percentage}
}
\value{
catchment
}
\description{
compute catchments
}
\details{
use habitat models for catchments
}
\author{
Barry Rowlingson
}

