% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceeds.R
\name{exceeds}
\alias{exceeds}
\title{Compute exceedences}
\usage{
exceeds(carmcout, t1 = attr(asamples, "start_date"), clevels = c(0.9, 0.95),
  elevels = c(2, 4, 8))
}
\arguments{
\item{carmcout}{carmc output}

\item{t1}{first date}

\item{clevels}{confidence interval levels}

\item{elevels}{exceedence thresholds}
}
\value{
a list of mu [nsites,ntimes] - sample mean; exceedence$e (exceedence levels):
exceedence$p[[nlevels]][nsites, ntimes] (exceedence probabilities),
cis[[nsites]][ntimes x ncilevels, [min,max,level,date]], dates
}
\description{
Compute exceedences
}
\details{
given an array of samples and a first date, compute confidence
intervals and exceedences probabilities of given levels and values
}
\author{
Barry Rowlingson
}

