% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.defaults}
\alias{mcmc.defaults}
\title{default carmc MCMC model parameters}
\usage{
mcmc.defaults(sigma2 = list(log.prior = use_dlnorm(-5, 3), init = exp(-5), h.l
  = 0.5, adjust = adjust_proposal), rho = list(init = 0.98, log.prior =
  use_dbeta(1, 1), h.t = 0.5, adjust = adjust_proposal), phi = list(init =
  0.7, log.prior = use_dbeta(1, 1), h.t = 0.5, adjust = adjust_proposal),
  beta = list(mean = function(p) {     rep(0, p) }, cov.inv = function(p) {   
   diag(1/1000, p) }), delta = list(values = seq(1, 100, 1), start_at = 10,
  step_size = 2))
}
\arguments{
\item{sigma2}{control for sigma-2}

\item{rho}{control for rho}

\item{phi}{control for phi}

\item{beta}{control for beta}

\item{delta}{control for delta}
}
\value{
a list of control parameters
}
\description{
Specification for carmc MCMC
}
\details{
MCMC model parameters
}
\author{
Barry Rowlingson
}

