% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_control.R
\name{mcmc_control}
\alias{mcmc_control}
\title{mcmc control parameters}
\usage{
mcmc_control(n.sim, burnin, thin, hthin, tmax = lubridate::dseconds(Inf))
}
\arguments{
\item{n.sim}{number of steps}

\item{burnin}{number of burn-in steps}

\item{thin}{sample thinning}

\item{hthin}{thinning of proposal parameters/acceptance sampling}

\item{tmax}{maximum run duration, not including burn-in.}
}
\value{
a list of parameters
}
\description{
mcmc control parameters
}
\details{
mcmc run control parameters
}
\author{
Barry Rowlingson
}

