% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchments.R
\name{multicatch}
\alias{multicatch}
\title{multiple catchments}
\usage{
multicatch(locs, pcs, idcolumn = 1)
}
\arguments{
\item{pcs}{vector of percentages}

\item{locations}{spatial points}
}
\value{
spatial polygons with percentage and source id as attribute
}
\description{
compute catchment at various percentages
}
\details{
given a few percentages, work out the catchment at each level
}
\author{
Barry Rowlingson
}

