% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencies.R
\name{voronoi_adjacency}
\alias{voronoi_adjacency}
\title{Find neighbours using Voronoi tessellation.}
\usage{
voronoi_adjacency(data, formula, scale = 1, PLOT = FALSE)
}
\arguments{
\item{data}{(data frame) with id and x, y columns}

\item{formula}{a formula of the form id~x+y}

\item{scale}{scale (divide) coordinates by this factor}

\item{PLOT}{(boolean) plot neighbour links when TRUE.}
}
\value{
list with components
  Adjacencies - adjacency matrix (TRUE means there is a neighbour).
  Distances - adjacency matrix with distances between neighbours.
  NumNeighbours - total number of neighbours.
}
\description{
Uses Voronoi tessellation to find nearest neighbours and their respective distances.  The neighbours are found using R package deldir.
}
\examples{

data=data.frame(i=1:6, x=c(1,2,1,3,4,8),y=c(1,1,5,5,4,8))
a=voronoi_adjacency(data,i~x+y)

}
\author{
Barry Rowlingson, Alison Hale
}

